// packages/contexts/ingestion/src/blockchains/bitcoin/types.ts
export interface BlockstreamTransaction {
  fee: number;
  locktime: number;
  size: number;
  status: {
    block_hash?: string;
    block_height?: number;
    block_time?: number;
    confirmed: boolean;
  };
  txid: string;
  version: number;
  vin: {
    is_coinbase: boolean;
    prevout: {
      scriptpubkey: string;
      scriptpubkey_address?: string;
      scriptpubkey_asm: string;
      scriptpubkey_type: string;
      value: number;
    };
    scriptsig: string;
    scriptsig_asm: string;
    sequence: number;
    txid: string;
    vout: number;
    witness: string[];
  }[];
  vout: {
    scriptpubkey: string;
    scriptpubkey_address?: string;
    scriptpubkey_asm: string;
    scriptpubkey_type: string;
    value: number;
  }[];
  weight: number;
}